% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artificial_loan_process}
\alias{artificial_loan_process}
\title{Loan Process Event Dataset}
\format{
A data frame consisting of 590 events:
\describe{
\item{case}{Trace indentifier}
\item{event}{Activity names}
\item{completeTime}{Timestamp when the event was completed}
}
}
\usage{
artificial_loan_process
}
\description{
An event dataset consisting the information for a loan process which consists of
several cases who performs different sequence of activites at different times
}
\keyword{datasets}
